% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateProcessingShareFBS.R
\name{calculateProcessingShareFBS}
\alias{calculateProcessingShareFBS}
\title{Function to compute Processing sharing: share that identify the quantity of primary availability 
that it is allocated in different productive processes.}
\usage{
calculateProcessingShareFBS(data, printSharesGraterThan1 = TRUE, param)
}
\arguments{
\item{data}{data table containing all the columns to compute processingSharing}

\item{printSharesGraterThan1}{logic parameter}

\item{params}{defaultProcessedItamParams parameters, object which contains the parameters}
}
\description{
Duplication used in Used in faoswsStandardization
}

