% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeYield.R
\name{computeYield}
\alias{computeYield}
\title{Function to compute and update yield}
\usage{
computeYield(
  data,
  processingParameters,
  formulaParameters,
  flagTable = ReadDatatable("flag_weight_table"),
  validFlags = ReadDatatable("valid_flags")
)
}
\arguments{
\item{data}{The data.table object containing the data.}

\item{processingParameters}{A list of the parameters for the production
processing algorithms.  See defaultProductionParameters() for a starting
point.}

\item{formulaParameters}{A list holding the names and parmater of formulas.
See \code{productionFormulaParameters}.}

\item{flagTable}{see data(faoswsFlagTable) in \pkg{faoswsFlag}. Please use
ReadDatatable("ocs2023_flagweight") for all future code.}

\item{validFlags}{Valid flags. Please use validFlags = ReadDatatable("valid_flags_ocs2023")
for future code}
}
\value{
The updated data.table.
}
\description{
Function to compute and update yield
}
