% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faoswsProduction-package.R
\docType{package}
\name{faoswsProduction-package}
\alias{faoswsProduction-package}
\alias{faoswsProduction}
\title{Package to perform the imputation of area harvested, production and
  yield for the FAO production domain.}
\description{
Repository to contain the code for imputing production data in the 
statistical working system.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWS-Methodology/faoswsProduction}
}

}
\author{
Joshua M. Browning \email{joshua.browning@fao.org}, Michael. C. J. 
  Kao \email{michael.kao@fao.org}
}
\keyword{package}
