% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllYieldKey.R
\name{getAllYieldKey}
\alias{getAllYieldKey}
\title{Function to get the complete yield key}
\usage{
getAllYieldKey(
  areaVar = "geographicAreaM49",
  itemVar = "measuredItemCPC",
  elementVar = "measuredElement",
  yearVar = "timePointYears",
  identityElements = c("31", "41", "51"),
  domain = "agriculture",
  dataset = "aproduction"
)
}
\arguments{
\item{areaVar}{The column name corresponding to the geographic area.}

\item{itemVar}{The column name corresponding to commodity item.}

\item{elementVar}{The column name corresponding to the measured element.}

\item{yearVar}{The column name corresponding to the time dimension}

\item{identityElements}{The elements codes which are related to the
production identity equation.}

\item{domain}{Domain from which to pull yield data}

\item{dataset}{Dataset from which to pull yield data}
}
\value{
A DatasetKey comprise of all the codes which are relevant to the
    production identity equation computation.
}
\description{
Function to get the complete yield key
}
