% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResult.R
\name{plotResult}
\alias{plotResult}
\title{This function has been created to plot PROCESSED data results
It is very Module-specific}
\usage{
plotResult(
  data,
  toPubblish,
  pathToSave,
  agDomain = "agriculture",
  agDataset = "aproduction",
  suaDomain = "suafbs",
  suaDataset = "sua"
)
}
\arguments{
\item{data}{The data.table we want to plot}

\item{toPubblish}{The set of commodities that we have to flag as "FAOSTAT" which are those that 
have a priority in the validation process.}

\item{pathToSave}{The path to the folder where the plots have to be saved.}

\item{agDomain}{Agriculture data domain. Used for naming data}

\item{agDataset}{Agriculture dataset name. Used for naming data}

\item{suaDomain}{SUA data domain. Used for naming data}

\item{suaDataset}{SUA dataset name. Used for naming data}
}
\description{
This function has been created to plot PROCESSED data results
It is very Module-specific
}
