% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResult.R
\name{plotResult}
\alias{plotResult}
\title{This function has been created to plot PROCESSED data results
It is very Module-specific}
\usage{
plotResult(data, toPubblish, pathToSave)
}
\arguments{
\item{data}{The data.table we want to plot}

\item{toPubblish}{The set of commodities that we have to flag as "FAOSTAT" which are those that 
have a priority in the validation process.}

\item{pathToSave}{The path to the folder where the plots have to be saved.}
}
\description{
This function has been created to plot PROCESSED data results
It is very Module-specific
}

